[BRIDGE]: Compat hooks for new-ioctl interface.
Replacement 64 bit compatibility code for the new ioctl's.  The new
ioctl's all pass through clean, but for the old style ioctl's it uses
the mis-feature of the earlier bridge-utils that they check the API version.

So if an old 32bit version of brctl is run on a 64bit platform it will
report
	bridge utilities not compatible with kernel version

Tested on Itanium 1; but should solve issue for sparc, ppc, and x86_64

BKrev: 40afbc45LtJYrwC5Yy0h_QG5Y4wqJw