Merge http://linux-mh.bkbits.net/bluetooth-2.6 into nuts.davemloft.net:/disk1/BK/net-2.6
2004/05/23 13:43:34+02:00 marcel
[Bluetooth] Define .kobj.k_name for the fake device

The PCMCIA devices are not devices for the kernel and the bt3c_cs
driver uses a fake device for calling request_firmware(). The fake
device initialization must also set .kobj.k_name to prevent an oops
until PCMCIA devices are fully integrated into the driver model.

2004/05/23 13:31:00+02:00 marcel
[Bluetooth] Use try_module_get() for RFCOMM sessions

It is not possible to use __module_get() when adding a new RFCOMM
session, because there is a case where no reference count is hold.
This happens when the module is not in use right now and an incoming
connection occurs.

2004/05/20 19:01:34+02:00 marcel
[Bluetooth] Kill duplicate includes

This patch removes the duplicate inclusion of linux/config.h in the
hci_uart driver.

Patch from Arthur Othieno <a.othieno@bluewin.ch>

BKrev: 40b12ffcGhm-2z9ws6nA7LPbaCmV8Q