Merge bk://bk.arm.linux.org.uk/linux-2.6-pcmcia into ppc970.osdl.org:/home/torvalds/v2.6/linux
2004/05/18 22:02:38+01:00 rmk
[PCMCIA] Convert IO resource allocation to use struct resource.

This causes PCMCIA to use struct resource internally for IO resources.
This means that we can keep track of the resource pointer, expand
this resource if necessary, and use release_resource() on it when
we're done.

This eventually means that we can change these to normal resources
which aren't marked busy.

2004/05/18 21:52:46+01:00 rmk
[PCMCIA] Fix a couple of resource bugs.

Fix resource database bug where base + num wraps to zero.  Also,
ensure that we always take the resource semaphore whenever we
allocate a resource.

BKrev: 40b163dbkrBtB1hZzvk3NNirDjNJNQ