[PATCH] ppc64: better stack traces
This improves the stack traces we get on PPC64 by putting a marker in
those stack frames that are created as a result of an interrupt or
exception.  The marker is "regshere" (0x7265677368657265).

With this, stack traces show where exceptions have occurred, which can
be very useful.  This also improves the accuracy of the trace because
the relevant return address can be in the link register at the time of
the exception rather than on the stack.  We now print the PC and
exception type for each exception frame, and then the link register if
appropriate as the next item in the trace.

BKrev: 40b2912fg1yuAj5tPNUD5xYnbRRRbw