Merge redhat.com:/spare/repo/netdev-2.6/pcnet32 into redhat.com:/spare/repo/net-drivers-2.6
2004/05/27 14:07:05-04:00 brazilnut
[PATCH] pcnet32: fix for patch 8 le16_to_cpu

Joe Perches pointed out an error in patch 8.  As I looked at the results
closer 3 out of 4 looked correct.  This fixes the output for the last
case.  Output is now the same on IA32 and PPC64 hardware.

2004/05/27 13:12:31-04:00 brazilnut
[PATCH] pcnet32: avoid timeout with tcpdump

This patch avoids a timeout when entering/exiting promiscuous mode with
tcpdump.  All pending xmit operations are flushed and if the queue is
full the timeout will occur.  The wake_queue avoids that timeout.
Tested ppc64 and ia32.

2004/05/27 13:12:23-04:00 brazilnut
[PATCH] pcnet32: correct printk for big-endian arch

This patch corrects a printk when executed on a big-endian system.  Tested
ppc64.

2004/05/27 13:12:15-04:00 brazilnut
[PATCH] pcnet32: fix bogus carrier errors with 79c973

The 79C973 version of the pcnet32 chipset would report all transmit
operations as being sent with a carrier error.  The root cause was
writing reserved bits in bcr33.  It caused the 973 to report xmit
as errors, and caused 975 and 976 to hard hang when reading bcr34.

My thanks to Bruce Penrod for his help in finding the cause of this
bug.

2004/05/27 13:12:06-04:00 brazilnut
[PATCH] pcnet32: limit frames received during interrupt.

The current code will allow an infinite number of frames to be received
during one interrupt.  Tested ia32 and ppc64.

2004/05/27 13:04:02-04:00 brazilnut
[PATCH] pcnet32: remove timer and complexity.

This removes the timer I recently added, and the 'ltint' interrupt
mitigation from the 79C971 and 79C972. It did not improve performance,
added complexity, and was a bad idea in hind sight.

2004/05/27 13:03:48-04:00 brazilnut
[PATCH] pcnet32: fix boundary comparison bug.

This patch fixes a boundary condition.  It is possible to get an
interrupt and have the transmit queue completely full.  Without this
the driver will leak skb's and pci maps (if the arch supports them).

2004/05/27 13:03:35-04:00 brazilnut
[PATCH] pcnet32: correct 79C976 variant string.

This patch correctly identifies the 79C976 as a PCnet/PRO not a FAST III.

2004/05/27 13:03:22-04:00 brazilnut
[PATCH] pcnet32: avoid hard hang with some chip variants.

This patch fixes a hard hang with 79C975 and 79C976 chip variants when
dumping BCR register #34.

2004/05/27 13:03:08-04:00 brazilnut
[PATCH] pcnet32: add static to two routines.

This patch adds 'static' to two routines.  The prototype has static but
the instantiation did not.

BKrev: 40b64979yV2rVq8GBHf-61VAqNLGqw