[PATCH] sparse: nf_sockopt() annotation
	netfilter set/getsockopt annotated.  That's a continuation of previous
patch - by now the only remaining place where setsockopt()/getsockopt()
userland pointers are not traced is vcc_[sg]etsockopt() and down into the
ATM drivers.  Everything from that area in net/* should be done by now.

	That's pretty much it as far as non-trivial splitting counts; remaining
patches in net/* are independent, tty and console stuff is also not hard to
split (not much of it, anyway) and what remains is arch-dependent code for
alpha and amd64.  Of course there'll be more stuff coming, but as far as I'm
concerned the rest of my pile is managable - and finally below 200Kb.  That'll
go tomorrow - right now I'm going down and getting some sleep...

BKrev: 40b9489dBfOjccjueXpOnNWMFzh94Q