Merge bk://gkernel.bkbits.net/libata-2.6 into ppc970.osdl.org:/home/torvalds/v2.6/linux
2004/06/02 17:51:11-04:00 arjanv
[libata] Use standard headers from include/scsi, not drivers/scsi

2004/06/02 17:31:53-04:00 jgarzik
[libata scsi] ack SYNCHRONIZE CACHE command

When taskfile delivery is improved, we will actually perform
this command, instead of no-oping it.

For now, this is cosmetic except for removing an annoying message
that some users see.

2004/06/02 17:29:00-04:00 jgarzik
[libata promise] revert broken taskfile delivery change

The ->tf_load and ->exec_command driver hooks were changed to assume
that PIO was the only type of taskfile ever delivered to these functions.
This will be true... in the future, but not today.  In other drivers
this change was not needed, but Promise executes commands differently
due to its "ATA packet" hardware features, so the Promise drivers need
this change reverted.

Diagnosis and initial fix by Brad Campbell <brad@wasp.net.au>

BKrev: 40beb18bcgVDf-PCEluLHIKgmRyRwA