[PATCH] Fix rivafb's OpenFirmware parsing
The attached patch fixes the EDID parsing for PPC on rivafb.  It actually
finds the EDID info in the OF Tree now.  I grabbed this from BenHs Tree as
of 2.6.5-rc3.  The current code has no chance to work since it doesn't walk
the device tree.  This helps rivafb on PPC at least a bit further...

Signed-off-by: Guido Guenther <agx@sigxcpu.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 40c209dajx-08nuw-JWem1mdE1BfDw